+++ packages/react-querybuilder/src/utils/objectKeys.ts
@@ -1,4 +1,5 @@
 /**
  * A typed proxy for `Object.keys`
  */
-export const objectKeys = <T extends object>(obj: T) => Object.keys(obj) as (keyof T)[];
+export const objectKeys = <T extends Record<string, any>>(obj: T) =>
+  Object.keys(obj) as (keyof T)[];